admin delete validation

+++ src/app/setting/userlist/userlist.component.ts
@@ -10,7 +10,6 @@ export interface userType{
   sortBy?:any,
   name?:any;
 }
-
 @Component({
   selector: 'app-userlist',
   templateUrl: './userlist.component.html',
@@ -21,7 +20,6 @@ export class UserlistComponent implements OnInit{
   dataComingStatus:boolean=false;
   // userDetails:{[key:string]:string}[]=[];
   userDetails:any=[];
-  successMessage!:string;
   reactiveForm!:FormGroup;
   editUserStatus:boolean=false;
   editRoleStatus:boolean=false;
@@ -38,7 +36,7 @@ export class UserlistComponent implements OnInit{
   sortBy:any='';
   name:any='';
   totalResult!: number;
-
+  adminId!:string;
 
   user:userType={
     limit:this.limit,
@@ -55,17 +53,19 @@ export class UserlistComponent implements OnInit{
       password:new FormControl(null,[Validators.required]),
       role:new FormControl(null,[ Validators.required])
     });
+
   }
   onGetUsers(user:userType){
+
     this.httpService.getUsers(user).subscribe((response:any)=>{
       console.log(response);
-      console.log(response.results.length);
-      // this.pageLength=response.results.length;
+      // console.log(response.results.length);
       this.totalResult=response.totalResults;
       this.pageCount=response.page;
       this.pageSize=response.totalPages;
       this.userDetails=response.results;
       this.dataComingStatus=true;
+
       this.pageCount==1?this.previousStatus=true:this.previousStatus=false;
       this.pageCount==this.pageSize?this.page1=this.page2=this.nextStatus=true:this.page1=this.page2=this.nextStatus=false;
       this.pageCount+1>=this.pageSize?this.page2=true:this.page2=false;
@@ -77,13 +77,24 @@ export class UserlistComponent implements OnInit{
     });
   }
   onDeleteUser(id:string){
-    this.httpService.deleteUser(id).subscribe(()=>{
-      // console.log('deleted successfully');
-      this.userService.showSuccess('Successfully Deleted');
-        this.router.navigate(['my-profile']);
+    this.httpService.getAdminId().subscribe((response:any)=>{
+      // console.log(response._id);
+      this.adminId=response._id;
     },err=>{
       this.httpService.error.next(err.error.message);
-      this.userService.showWarning(err.error.message);
+    },()=>{
+      if(this.adminId!=id){
+        this.httpService.deleteUser(id).subscribe(()=>{
+          this.userService.showSuccess('Successfully Deleted');
+            this.router.navigate(['my-profile']);
+        },err=>{
+          this.httpService.error.next(err.error.message);
+          this.userService.showWarning(err.error.message);
+        });
+      }
+      else{
+        this.userService.showWarning('Admin can not be deleted');
+      }
     });
   }
   editUser(user:object){
@@ -98,7 +109,6 @@ export class UserlistComponent implements OnInit{
       role:this.specificUser.role
     });
     this.reactiveForm.controls['role'].disable();
-
   }
   editRole(user:object){
     this.editUserStatus=false;
@@ -124,7 +134,6 @@ export class UserlistComponent implements OnInit{
       this.httpService.updateUserRole(id,{role}).subscribe((res)=>{
         console.log(res);
         this.userService.showSuccess('Role updated successfully');
-
         setTimeout(()=>{
           this.reactiveForm.reset();
           this.router.navigate(['my-profile']);
@@ -136,7 +145,6 @@ export class UserlistComponent implements OnInit{
     }
     else{
       this.httpService.updateUserDetails(id,{email,password,name}).subscribe(()=>{
-        // console.log('user updated');
         this.userService.showSuccess('User updated successfully');
 
         setTimeout(()=>{
@@ -149,25 +157,6 @@ export class UserlistComponent implements OnInit{
       });
     }
   }
-  // pegination 
-  // onPrevious(){
-  //   if(this.pageCount>1){
-  //     this.pageCount=this.pageCount-1;
-  //     this.onGetUsers({...this.user,page:this.pageCount});
-  //   }
-    
-  // }
-  // onNext(){
-  //   if(this.pageCount+2<this.pageSize){
-  //     this.pageCount=this.pageCount+1;
-  //     this.onGetUsers({...this.user,page:this.pageCount});
-  //   }
-  // }
-  // onPage(pageCount:any){
-  //   this.user.limit=this.limit;
-  //   this.user.page=pageCount;
-  //   this.onchange();
-  // }
   onChangeSortBy(){
     if(this.sortBy!=''){
       this.user.sortBy=this.sortBy;
@@ -197,7 +186,6 @@ export class UserlistComponent implements OnInit{
   onPageChange(pageCount:any){
     this.pageCount=pageCount;
     this.onGetUsers({...this.user,page:this.pageCount});
-
   }
   onchange(){
     this.dataComingStatus=false;
